Move pending claim teardown contract to worker channel
